backend/requirements.txt
fastapi==0.116.1
uvicorn==0.32.1
pydantic==1.10.18
python-multipart==0.0.12
requests==2.32.3
python-chess==1.999
psycopg2-binary==2.9.9
redis==5.2.0
python-dotenv==1.0.1